Introduce finalizer skeleton
Now we are able to tell if the object is to be deleted and run actions based on the outcome. We will add kube finalizer and requeue if the deletion timestamp is set. This will tell kube API to not delete the object before we are done with our finalization logic.